+++ src/nitpick/project.py
@@ -242,7 +242,7 @@ class Project:
 
     def create_configuration(self) -> None:
         """Create a configuration file."""
-        from nitpick.style import Style  # pylint: disable=import-outside-toplevel
+        from nitpick.style.style import Style  # pylint: disable=import-outside-toplevel
 
         doc = document()
         doc.add(comment("This file was generated by the `nitpick init` command"))
